analyzer: get the element:pad names to put in the results
Makes it more meaningful and easier to read. Also remove the peer-pad
from our logs as the query is done on a single pad and is easier to
read it that way. Sometimes the query isn't done from one pad to another
but just on a single pad (elements querying their own pads) and it makes
it hard to differentiate when one element is querying other element's pads
or when it is doing it by itself.